refactor: namespaces for test initializers and fixtures
Standardize and correct namespace usage across test and initializer files. Move idempotency and outbox initializers into dedicated sub-namespaces (`Internals.Idempotency`, `Internals.Outbox`). Update test files to reference new namespaces and correct using directives, improving code organization and maintainability.
