+++ tests/NetEvolve.Pulse.Tests.Integration/Idempotency/InMemoryEntityFrameworkIdempotencyTests.cs
@@ -1,7 +1,9 @@
-namespace NetEvolve.Pulse.Tests.Integration.Idempotency;
+﻿namespace NetEvolve.Pulse.Tests.Integration.Idempotency;
 
 using NetEvolve.Extensions.TUnit;
 using NetEvolve.Pulse.Tests.Integration.Internals;
+using NetEvolve.Pulse.Tests.Integration.Internals.Idempotency;
+using NetEvolve.Pulse.Tests.Integration.Internals.Services;
 
 [ClassDataSource<InMemoryDatabaseServiceFixture, EntityFrameworkIdempotencyInitializer>(
     Shared = [SharedType.None, SharedType.PerTestSession]
