+++ tests/NetEvolve.Pulse.Tests.Integration/Idempotency/PostgreSqlEntityFrameworkIdempotencyTests.cs
@@ -1,7 +1,8 @@
-namespace NetEvolve.Pulse.Tests.Integration.Idempotency;
+﻿namespace NetEvolve.Pulse.Tests.Integration.Idempotency;
 
 using NetEvolve.Extensions.TUnit;
 using NetEvolve.Pulse.Tests.Integration.Internals;
+using NetEvolve.Pulse.Tests.Integration.Internals.Idempotency;
 
 [ClassDataSource<PostgreSqlDatabaseServiceFixture, EntityFrameworkIdempotencyInitializer>(
     Shared = [SharedType.None, SharedType.None]
