+++ tests/NetEvolve.Pulse.Tests.Integration/Internals/Idempotency/RedisIdempotencyInitializer.cs
@@ -1,4 +1,4 @@
-namespace NetEvolve.Pulse.Tests.Integration.Internals;
+﻿namespace NetEvolve.Pulse.Tests.Integration.Internals.Idempotency;
 
 using Microsoft.Extensions.DependencyInjection;
 using NetEvolve.Pulse.Extensibility;
