+++ tests/NetEvolve.Pulse.Tests.Integration/Internals/Services/RedisContainerFixture.cs
