+++ tests/NetEvolve.Pulse.Tests.Integration/Internals/Services/RedisServiceFixture.cs
