+++ tests/NetEvolve.Pulse.Tests.Integration/Internals/Services/SQLiteDatabaseServiceFixture.cs
