+++ tests/NetEvolve.Pulse.Tests.Integration/Internals/Services/SqlServerDatabaseServiceFixture.cs
