+++ tests/NetEvolve.Pulse.Tests.Integration/Outbox/MongoDbOutboxTests.cs
@@ -1,7 +1,9 @@
-namespace NetEvolve.Pulse.Tests.Integration.Outbox;
+﻿namespace NetEvolve.Pulse.Tests.Integration.Outbox;
 
 using NetEvolve.Extensions.TUnit;
 using NetEvolve.Pulse.Tests.Integration.Internals;
+using NetEvolve.Pulse.Tests.Integration.Internals.Outbox;
+using NetEvolve.Pulse.Tests.Integration.Internals.Services;
 
 [ClassDataSource<MongoDbDatabaseServiceFixture, MongoDbOutboxInitializer>(Shared = [SharedType.None, SharedType.None])]
 [TestGroup("MongoDB")]
